Fix socket error for slow devices
Similar to EINTR recv() will return EAGAIN if _no_ data is available. 

This fixes a case where the device is to slow to send data in time and helps to drain the output buffer on the device to keep the connection in sync.

The risk for breakage is low as there is a timeout in place.